assembler: start reorganizing
