+++ assembler/.gitignore
@@ -2,3 +2,6 @@
 **/*.rs.bk
 Cargo.lock
 *.iml
+*.asm
+*.obj
+*.mem
